Refactor subtitlecat service into modules
