+++ a4kSubtitles/services/subtitlecat/__init__.py
@@ -1,0 +1,7 @@
+from .request import build_search_requests, parse_search_response, build_download_request
+
+__all__ = [
+    'build_search_requests',
+    'parse_search_response',
+    'build_download_request',
+]
